move serializer to dom
